Add some tests to the s3proxy.
Several kind of tests are available (check README for more details) :

- Unit tests

Unit tests are used to verify the wanted behaviour of the s3proxy.
They don't need any external components, it is using a fake S3 backend for running tests.

- Integration tests

Integration tests are used to verify the integration with a real s3 backend and a rsyslog server.
In our tests we are using minio server which provides a S3 compatible API.

- End-to-end tests

End-to-end tests are used to verify the whole integration with a s3proxy running standalone in its container, a S3 backend (minio) and a rsyslog.
The integration tests simulate external calls from a program using the s3proxy.