Get rid of the s3proxy syslog mtype in favor of the generic mtype
In this PR https://github.com/mirakl/production/pull/4325
I added a `generic` mtype which simply routes the logs to
`${namespace}/${mserver}.log`.

This PR changes s3proxy to make use of thar routing rule instead of the
`s3proxy` specific one so that we can cleanup the logstash config.