Rework the CD process
Currently the build & push image is managed by makefile. User should run
make command to build and push is our personnal repository.
We don't want it, and use dockerhub as default docker registry for s3proxy.

To deal with that:
 - Fix end2end integration test: this process should use an image
already builded pushed and available. I think it doesn't make sense
cause we have to run test before to push the final image. So at this
purpose use the "build" image to run test.
 - Remove VERSION variable used in Dockerfile: dockerhub launch directly
   the build with Dockerfile, which contains VERSION variable. BUT the
VERSION is never filled on dockerhub. So remove this VERSION on
dockerfile and use the VERSION provided in Makefile (which is launched
in Dockerfile)

Now if we need to build and push the new version on docker hub, just
create a tag on master should be enough.